add odd dim check without subsampling back
I was assuming if subsampled planes are even then the full res must be even too, while that mostly true, the integer division will make it not work pretty well.
Say for example 321/2=160, although the odd dimension + subsample probably won't work for other parts of the script anyway, it's better to add it back.